Stray file was moved, fixing
